8_official_a2a_protocol/requirements.txt
a2a-sdk
anthropic
google-genai
python-dotenv
uvicorn
httpx
